+++ dd-java-agent/agent-iast/src/main/java/com/datadog/iast/model/json/VulnerabilityEncoding.java
@@ -1,7 +1,6 @@
 package com.datadog.iast.model.json;
 
 import com.datadog.iast.model.VulnerabilityBatch;
-import com.datadog.iast.model.json.TruncatedVulnerabilitiesAdapter.TruncatedVulnerabilities;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 
@@ -13,6 +12,9 @@ public class VulnerabilityEncoding {
   private static final JsonAdapter<VulnerabilityBatch> BATCH_ADAPTER =
       MOSHI.adapter(VulnerabilityBatch.class);
 
+  private static final JsonAdapter<TruncatedVulnerabilities> TRUNCATED_VULNERABILITIES_ADAPTER =
+      MOSHI.adapter(TruncatedVulnerabilities.class);
+
   public static String toJson(final VulnerabilityBatch value) {
     String json = BATCH_ADAPTER.toJson(value);
     return json.getBytes().length > 25000
@@ -22,6 +24,6 @@ public class VulnerabilityEncoding {
 
   static String getExceededTagSizeJson(final TruncatedVulnerabilities truncatedVulnerabilities) {
     // TODO report via telemetry
-    return MOSHI.adapter(TruncatedVulnerabilities.class).toJson(truncatedVulnerabilities);
+    return TRUNCATED_VULNERABILITIES_ADAPTER.toJson(truncatedVulnerabilities);
   }
 }
