+++ dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/model/json/VulnerabilityEncodingTest.groovy
@@ -39,7 +39,7 @@ class VulnerabilityEncodingTest extends DDSpecification {
     }''', result, true)
 
     when:
-    result = VulnerabilityEncoding.getExceededTagSizeJson(new TruncatedVulnerabilitiesAdapter.TruncatedVulnerabilities(value.getVulnerabilities()))
+    result = VulnerabilityEncoding.getExceededTagSizeJson(new TruncatedVulnerabilities(value.getVulnerabilities()))
 
     then:
     JSONAssert.assertEquals('''{
@@ -445,7 +445,7 @@ class VulnerabilityEncodingTest extends DDSpecification {
       ))
 
     when:
-    final result = VulnerabilityEncoding.getExceededTagSizeJson(new TruncatedVulnerabilitiesAdapter.TruncatedVulnerabilities(value.getVulnerabilities()))
+    final result = VulnerabilityEncoding.getExceededTagSizeJson(new TruncatedVulnerabilities(value.getVulnerabilities()))
 
     then:
     JSONAssert.assertEquals('''{
@@ -485,7 +485,7 @@ class VulnerabilityEncodingTest extends DDSpecification {
       ))
 
     when:
-    final result = VulnerabilityEncoding.getExceededTagSizeJson(new TruncatedVulnerabilitiesAdapter.TruncatedVulnerabilities(value.getVulnerabilities()))
+    final result = VulnerabilityEncoding.getExceededTagSizeJson(new TruncatedVulnerabilities(value.getVulnerabilities()))
 
     then:
     JSONAssert.assertEquals('''{
